Working Bepsoke Attachments
* Added Lots, honestly didn't update in too long so there are hundreds of changes
* Added Potatoes BGM loader, it breaks if loaded in
* Added Per Round costs for magazine upgrades

+++ Packer-SupplyRaid/src/Plugin/src/Scripts/H3MP/SR_Networking.cs
@@ -298,7 +298,7 @@ namespace SupplyRaid
         //Game Options Received
         void GameOptions_Handler(int clientID, Packet packet)
         {
-            int optionPlayerCount = packet.ReadInt();
+            float optionPlayerCount = packet.ReadInt();
             float optionDifficulty = packet.ReadFloat();
             bool optionFreeBuyMenu = packet.ReadBool();
             bool optionSpawnLocking = packet.ReadBool();
